requirements.txt
python-dotenv
twitchio
PyDeepLX
langdetect
